==> The_project/1. Happiness report web scraping 2018.ipynb <==
### 1. Importing needed libraries

import requests
import pandas as pd
from bs4 import BeautifulSoup
import os

### 2. Calling on url and see layout

url = "https://en.wikipedia.org/wiki/World_Happiness_Report"
html = requests.get(url).content
soup = BeautifulSoup(html, "lxml")


### 3. Selecting correct table 

table = soup.find_all('table',{'class':'wikitable sortable'})[0]

### 4. Checking on rows, layout, data and selecting needed info

rows = table.find_all('tr')
#rows = [row.text.strip().split("\n") for row in rows]
rows = [row.text.replace("\n\n", ",").replace("\n", ",").replace("\xa0", "").strip(",").split(",") for row in rows]
rows [:2]

### 5. Creating the dataframe

colnames = rows[0]
data = rows[1:]

happiness_2018_extended = pd.DataFrame(data, columns=colnames)
happiness_2018_extended.head(20)


### 6. Basic data cleaning + data selecting

os.getcwd()
os.chdir("../Data")

#####  saving original to csv (to avoid data loss)

happiness_2018_extended.to_csv("backup_raw_data/happiness_2018_extended.csv")

##### selecting countries for research

# countries selected for use:
# spain, netherlands, france, denmak, sweden, italy, poland, belgium, greece, austria

country_list = ["Spain", "Netherlands", "France", "Greece", "Austria", "Denmark", "Belgium", "Italy", "Poland", "Sweden"]

happiness_2018 = pd.read_csv("../Data/backup_raw_data/happiness_2018_extended.csv")
happiness_2018 = happiness_2018[(happiness_2018["Country or region"].isin(country_list))]

happiness_2018 = happiness_2018.rename(columns={"Country or region":"Country",
                                                "Happiness score":"Score",
                                                "Generosity":"Generosity importance", 
                                                "GDP per capita":"GDP", 
                                                "Freedom to make life choices":"Freedom"})
# no need to have the rank and the unnamed as we will not analyse by rank but by score:
happiness_2018 = happiness_2018.drop(["Overall rank"], axis=1)
happiness_2018 = happiness_2018.drop(["Unnamed: 0"], axis=1)

happiness_2018

##### save new to csv to be able to work on it

happiness_2018.to_csv("../Data/happiness_2018.csv")
