==> The_project/4. Database_creation_seeding.ipynb <==
### Creating connection with the project's database and creating the tables in it

from sqlalchemy import create_engine
import pymysql
import pandas as pd
import numpy as np

driver = "mysql+pymysql:"
user = "root"
password = "ironhack"
ip = "35.188.185.234"
database = "Project_3"

connection_string = f"{driver}//{user}:{password}@{ip}/{database}"
engine = create_engine(connection_string)

### Inputting data into our database and checking operation success

countries = pd.read_csv("../Data/countries.csv")

countries

countries.to_sql("country", con=engine, if_exists="append", index=False)

engine.execute("SELECT * FROM country").fetchall()

elections = pd.read_csv("../Data/Elections_10_countries.csv")

elections.head()

elections.to_sql("elections", con=engine, if_exists="append", index=False)

# To check whether all data has been input correctly into our database
engine.execute("SELECT * FROM elections").fetchall()

engine.execute("SELECT * FROM happiness").fetchall()

happiness_all = pd.read_csv("../Data/happiness_all.csv")

happiness_all.to_sql("happiness", con=engine, if_exists="append", index=False)
